requirements.txt
pip~=23.2.1
wheel~=0.41.2
setuptools~=68.2.0
google-cloud-storage
apache_beam
google-cloud-firestore